refactor: 상마 포진 반환 객체 타입으로 변경
