+++ src/main/java/domain/board/formation/InitialFormationType.java
